# feat: optimize multi-agent tool sharing with LRU cache