--- a/src/crewai/crew.py
+++ b/src/crewai/crew.py
@@ -22,6 +22,10 @@ from opentelemetry import baggage
 from opentelemetry.context import attach, detach
 
 from crewai.utilities.crew.models import CrewContext
+from crewai.utilities.tool_sharing_cache import (
+    get_tool_sharing_cache,
+    should_use_tool_sharing,
+)
 
 from pydantic import (
     UUID4,
@@ -923,6 +927,42 @@ class Crew(FlowTrackable, BaseModel):
     def _prepare_tools(
         self, agent: BaseAgent, task: Task, tools: Union[List[Tool], List[BaseTool]]
     ) -> List[BaseTool]:
+        # Check if tool sharing cache can be used
+        if should_use_tool_sharing(tools):
+            # Extract agent capabilities for cache key generation
+            allow_delegation = hasattr(agent, "allow_delegation") and getattr(
+                agent, "allow_delegation", False
+            )
+            allow_code_execution = hasattr(agent, "allow_code_execution") and getattr(
+                agent, "allow_code_execution", False
+            )
+            multimodal = hasattr(agent, "multimodal") and getattr(
+                agent, "multimodal", False
+            )
+            process_type = str(self.process)
+
+            # Try to get tools from cache
+            tool_cache = get_tool_sharing_cache()
+            cached_tools = tool_cache.get_tools(
+                agent_id=str(agent.id)
+                if hasattr(agent, "id")
+                else str(hash(agent.role)),
+                task_id=str(task.id)
+                if hasattr(task, "id")
+                else str(hash(task.description)),
+                tools=tools,
+                allow_delegation=allow_delegation,
+                allow_code_execution=allow_code_execution,
+                multimodal=multimodal,
+                process_type=process_type,
+            )
+
+            if cached_tools is not None:
+                return cached_tools
+
+        # Original tool preparation logic if not cached
+        original_tools = tools
+
         # Add delegation tools if agent allows delegation
         if hasattr(agent, "allow_delegation") and getattr(
             agent, "allow_delegation", False
@@ -951,8 +991,38 @@ class Crew(FlowTrackable, BaseModel):
         ):
             tools = self._add_multimodal_tools(agent, tools)
 
-        # Return a List[BaseTool] which is compatible with both Task.execute_sync and Task.execute_async
-        return cast(List[BaseTool], tools)
+        # Cast to List[BaseTool] for compatibility
+        prepared_tools = cast(List[BaseTool], tools)
+
+        # Cache the prepared tools if tool sharing is beneficial
+        if should_use_tool_sharing(original_tools):
+            allow_delegation = hasattr(agent, "allow_delegation") and getattr(
+                agent, "allow_delegation", False
+            )
+            allow_code_execution = hasattr(agent, "allow_code_execution") and getattr(
+                agent, "allow_code_execution", False
+            )
+            multimodal = hasattr(agent, "multimodal") and getattr(
+                agent, "multimodal", False
+            )
+            process_type = str(self.process)
+
+            tool_cache.store_tools(
+                agent_id=str(agent.id)
+                if hasattr(agent, "id")
+                else str(hash(agent.role)),
+                task_id=str(task.id)
+                if hasattr(task, "id")
+                else str(hash(task.description)),
+                tools=original_tools,
+                prepared_tools=prepared_tools,
+                allow_delegation=allow_delegation,
+                allow_code_execution=allow_code_execution,
+                multimodal=multimodal,
+                process_type=process_type,
+            )
+
+        return prepared_tools
 
     def _get_agent_to_use(self, task: Task) -> Optional[BaseAgent]:
         if self.process == Process.hierarchical: