@chi2liu 2025-08-08
> I have another question for this cache system
> 
> What would happen in this case?
> 
> ```
> read_file_chunk_tool = FileReadTool(file_path="file-chunk.txt")
> read_file_knowledge_tool = FileReadTool(file_path="knowledge.txt")
> 
> Agent(.., tools=[read_file_chunk_tool, read_file_knowledge_tool])
> ```
> 
> Will those tool instances be treated individually? I'm a bit concerned because we're currently using only the tool name as the cache key.

You're absolutely right to raise this concern! The original implementation indeed had this issue - using only tool names would have caused collisions between different `FileReadTool` instances with different `file_path` values.

Thanks to your review, this has been fixed in the updated implementation.